requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
apptools==5.3.0
asttokens==3.0.0
attrs==23.2.0
backcall==0.2.0
black==24.4.2
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
ccimport==0.4.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.8
cumm-cu114==0.4.11
cupy-cuda113==10.6.0
cycler==0.12.1
Cython==3.0.10
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==4.4.2
descartes==1.1.0
docker-pycreds==0.4.0
dpcpp-cpp-rt==2025.1.0
einops==0.8.0
entrypoints==0.4
envisage==7.0.3
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
fastrlock==0.8.2
filelock==3.14.0
fire==0.6.0
flake8==7.1.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.2.1
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.64.1
huggingface-hub==0.30.1
Hydra==2.5
hydra-core==1.3.2
idna==3.4
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
intel-cmplr-lib-rt==2025.1.0
intel-cmplr-lib-ur==2025.1.0
intel-cmplr-lic-rt==2025.1.0
intel-opencl-rt==2025.1.0
intel-openmp==2025.1.0
intel-sycl-rt==2025.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lark==1.1.9
lazy_loader==0.4
llvmlite==0.41.1
-e git+https://github.com/YkiWu/EmbodiedOcc.git@5443e5d37fa09adf895f8237f6460daf95a4c688#egg=local_aggregate&subdirectory=model/head/gaussian_occ_head/ops/localagg
lyft-dataset-sdk==0.0.8
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mayavi==4.7.3
mccabe==0.7.0
mdurl==0.1.2
mkl==2025.1.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.0.1
mmdet==3.0.0
mmdet3d==1.1.1
mmengine==0.10.4
mmsegmentation==1.2.2
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.33.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.23.2
nuscenes-devkit==1.1.11
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pccm==0.4.11
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
plyfile==1.0.3
portalocker==2.10.1
prettytable==3.10.2
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.1
pycocotools==2.0.7
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pyface==8.0.0
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.1.2
PyQt5_sip==12.15.0
pyquaternion==0.9.9
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
pytorch3d==0.7.5
pytz==2023.4
PyVirtualDisplay==3.0
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.35.1
regex==2024.7.24
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.19.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.13.0
setproctitle==1.3.3
Shapely==1.8.5.post1
six==1.17.0
smmap==5.0.1
spconv-cu114==2.3.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.3.0
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.7
tomli==2.0.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
trimesh==4.4.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.152
ultralytics-thop==2.0.14
umf==0.10.0
urllib3==1.26.19
vtk==9.0.1
wandb==0.13.9
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
yapf==0.40.2
zipp==3.19.2
